fix AntiLink detection and canonical wallpaper command
