+++ tests/moderation-and-vv.test.js
@@ -1,6 +1,7 @@
 const test = require('node:test');
 const assert = require('node:assert/strict');
 const { isForwardedMessage } = require('../src/Commands/Admin/antiforward');
+const antilink = require('../src/Commands/Admin/antilink');
 const vv = require('../src/Commands/Converter/view-once');
 
 test('AntiForward detects forwarding metadata in every message container', () => {
@@ -9,12 +10,19 @@ test('AntiForward detects forwarding metadata in every message container', () =>
   assert.equal(isForwardedMessage({ message: { conversation: 'ordinary text' } }), false);
 });
 
+test('AntiLink detects and extracts TikTok short links from nested message text', () => {
+  const message = { extendedTextMessage: { text: 'https://vt.tiktok.com/ZSVAh671Y/' } };
+  const text = antilink.getMessageText(message).join(' ');
+  assert.equal(antilink.hasLink(text), true);
+  assert.deepEqual(antilink.extractUrls(text), ['https://vt.tiktok.com/ZSVAh671Y/']);
+});
+
 test('view-once module exposes the automatic forwarding hook', () => {
   assert.equal(typeof vv.handleAutoVV, 'function');
   assert.ok(vv.alias.includes('autovv'));
 });
 
 test('wallpaper command loads without an undefined prefix reference', () => {
-  const wallpaper = require('../src/Commands/Search/wp');
+  const wallpaper = require('../src/Commands/Search/WP');
   assert.equal(wallpaper.usage, '.wallpaper <query>');
 });
